requirements.txt
numpy~=2.1.3
pymunk~=6.11.0
numba~=0.61.0
matplotlib~=3.10.0
scipy~=1.15.1
pyglet~=2.1.2
pillow~=11.1.0